requirements.txt
django
django-bootstrap-v5
django-crispy-forms
channels
bootstrap4
dpd_static_support
whitenoise
passlib
djangorestframework
django-filter
